Add support for complex secrets
This commit refactors how secrets are configured and allows users to
supply complex values (instead of just simple strings). This allows us
to encode secrets like "username/password" into a secret and not require
seperate embeddings in referrers.

As part of this change, I've removed a bunch of the configuration
validation that was in place originally. We can revisit stronger config
validation once we have some experience working with Lighthouse configs.