Decouple the sqlsync package from systems
We will want to re-use this libraries and stacks so do not couple to
systems. It should be fine to sprintf the FROM and primary key name into
the SQL query because those are hardcoded into the service. Do not
sprintf the id value however as that could open up to injections.