internal/s3: Document the AWS, GCP, Azure authentication processes.
If no credentials are provided, default to the SDK preferred default
credential chains. The default behavior can be restricted with
explicit configuration to a particular credential discovery mechanism.

For now the metadata service provided credentials are something you
cannot explicitly configure for but to use them you have to rely on
the default chains. The assumption here is the common case is *not* to
configure any authentication/credentials when using any object
storage.