Add initial version of backtest tool
Pretty output is not supported yet and there is no logic to check if the
bundle in object storage differs from the bundle that generated the
decision--so if there were policy changes that occurred after the
decision was generated, differences could be expected. We can improve
the tool to run the decision log through other DAS generated bundles to
get a better sense of correctness.