Add new migration tool
The migration tool will create systems and libraries like the existing
export tool however it will also create stacks (along with the necessary
library to hold stack rego) and set migrate the system and stack labels
and selectors accordingly. The migration tool also operates differently
with respect to copying non-git backed files out of the tenant--instead
of relying on bundle contents, the files to copy for systems and stacks
are pulled from the system and stack resources respectively. This avoids
inclusion of boilerplate generated by DAS as well as confusion between
which files belong in the system vs. stack vs. library.