libraries: remove colon from envoy-2.1 conflicts library path
It seems the colon in the filename stops the embed package from
including the subdirectories properly. Removing the colon solved this.
Note, this change means the file will not show up in the bundle in the
same location as old bundles built by DAS. This will make bundle
comparisons more complicated. For now, we can remove the colon to make
progress but we may want to revisit ths and have some way to escape
colons in directory names in the library assets.