cmd/migrate: add initial support for datasource migration
Currently only the "rest/push" datasource from DAS is supported. Next up
we can add the "http/pull" datasource. The --datasource flag on the
migrate command was added because we are not yet sure if we want to
recommend that users export data from DAS (as they will need to make
sure whatever services are pointing at and pushing data get updated to
point at the new API.