internal/builder: add support for include paths on git sync
This commit adds support for an include path filter on git sync
configurations. This lets users point git sync at a remote repo and only
include paths that match the filter (relative to the git_root[/path]
prefix). This is needed to support the old workspace git configuration
convention that required a stacks/stackid directory layout.