e2e: disable backtest time inflation check
The e2e test cannot produce reliable timing right now because we are
looking at the logged decision. On GHA the latency can vary wildly so
disable for now. We can return to this and improve the backtest to
re-evaluate the decision locally with the original bundle in order to
produce more accurate results.