internal/database: remove join for git secrets
Bundles no longer refer to git repos so there are no git secrets in the
bundles_secrets join-table. This commit just removes the logic that
checked for git secrets and renames the function accordingly. I'm not
sure if we support object storage w/o credentials however this was the
original logic so I'm leaving it intact.