internal: add token authentication
This commit adds tokens to tthe configuration model and implements
bearer token authentication in the HTTP server. Internally, all access
is made via principals that are mapped into by the authentication
middleware. For now we only support a single type of principal (API
token) but in the future we could add others.