internal/server: add initial implementation of GET v1/sources
This commit adds an endpoint on the server to list sources. As part of
this change the authz package is updated to export the expression and
column refs types so that they can be referenced outside by the
server/database/service packages.

On startup, the service will assert a principal with administrator
access and all internal operations will eventually be performed in this
context.