cmd/migrate: fine-grained config splitting
This commit should make it easier for users to test migration on a
series of systems. Instead of writing out a fixed set of config files,
we generate one for each:

* system containing the bundle and source for that system
* stack containing the stack and source for that stack
* library containing the source for that library

And then separate files foreach additional required source and secret.
This way the migrate command can be run multiple times to build up a
config directory that can be unioned together.