authz_test: make tests repeatable, fix flaky TestPartialStringArgs
TestPartialStringArgs was misleading a bit: First, it used the cache, so running
it multiple times wouldn't evaluate the nondeterministic code path multiple times
but kept returning the cached result. Secondly, the expected query contains four
parts, the last two of which are equal. This is because the partial eval result
queries contain (almost) the same expressions, one for viewer, one for admin, and
the don't come out of PE in a stable order.

So now, we'll check that the first 6 args match exactly, and assert that one of
the two (equivalent) versions of the last 4 args is returned.

Also fixes some more stuff to make it possible to run all the authz tests multiple
times.
Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>