http: expose secrets CRUD (#101)
* database: add {Get,List,Delete}Secret(s)

To support listing them, we're adding an auto-incrementing id column
to the table.

Let's figure out some time in the future when we stop doing such moves,
and introduce a migrations framework of some sort.

Also adding rows.Err() checks that are required by the driver, but had
been missing in a few places.

* server: expose secrets CRUD via handlers, with List and Get result as SecretRef

This way, the API will never hand out a secret. It'll only let you
know the names of secrets it knows, and let you delete or update them
by name.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>