Fix failing cli builds with --non-interactive flag exiting with success (#98)
os.Exit(1) was only called within `printReport` which is only executed
when `--non-interactive` is not used. This lead to failing bundle builds
still resulting in exit code 0 when the flag was used.

After this commit the build results are collected regardless of the
`--non-interactive` flag and the exit code is set properly based on
those results.

Signed-off-by: Stefan Lau <stefan.lau@zeiss.com>