builder: apply excludes when checking bundle roots (#110)
* builder: apply excludes when checking bundle roots

Previously, the per-source excludes hadn't been applied when walking
data and policy files to figure out the roots contributions of a
source.

Now, the same FilterFS that is used for building the bundle is used
to determine its roots, yielding more consistent results.

Fixes #107.

In the e2e tests, I have added something to build_git_local.txtar,
because that one more closely mirrors the problem described in #107.
For the directory/paths setup, it isn't as easy to reproduce it,
since we have to declare included paths. For git, the build process
will pull in _all_ files, if they haven't been excluded.


Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>